feat(dracut.sh): save configuration of image in images
Save the configuration files used to build an image in
the initrd image and incorporate them in --rebuild mode.
This facilitates image rebuilding and supplements the
build-parameter.txt record of the dracut command line
arguments.

Allow the --rebuild [output file] to be optional, as
documented. Update and fix the documentation of --rebuild,
which is hidden on some online versions of the manual page.

Also, drop a redundant existence check for config files
in the configuration directory.