fix(network-manager): start the online hooks from nm-dispatcher
So there are three steps we need to do when a NIC appears:
1) We want udev to do its stuff
2) We want NetworkManager to setup the interface
3) We want to run the online hooks

On machines without systemd this works just fine. nm-run is called after udev settles and
it starts NM in the oneshoot mode and call the hooks. Everything is done nicely synchronously.

On machines with systemd, NM is started as a daemon earlier and listens to udev events.
When it gets the event that device appeared it starts the configuration.

But we only run the hooks once after NM told us that it finished the setting up everything.
But if an additional device appears later, we will not start the related hooks again.

So let's use NM's dispatcher to start the hooks.