fix(memstrack): change memstrack report hook during kdump
Previously memstrack report is attached to the cleanup hook. It is OK
if memstrack is enabled for the 1st kernel. However if we want to
memstrack the 2nd kernel during kdump process, the cleanup hook may not
get triggered before the kdump finishes vmcore dumping and forcing reboot.
As a result, the log collected by memstrack analyze service will not be
outputted.

For example in fedora, the kdump-capture.service is wanted by
initrd.target, same as memstrack.service. After the successful vmcore
dumping, it will force reboot by 'reboot -f' by default [1], which will
not wait for cleanup hook to finish.

In this patch, we will replace memstrack report cleanup hook with
memstrack-report.service in kdump case, and make it wanted by initrd.target.
So memstrack report will be triggered earlier.

[1]: https://src.fedoraproject.org/rpms/kexec-tools/blob/rawhide/f/dracut-kdump.sh#_29

Signed-off-by: Tao Liu <ltao@redhat.com>