fix(debug xtrace): failures with mksh and with rd.live.debug
Drop debug_on() debug_off() project-wide as setting xtrace is only
local in mksh.  Use set -x directly instead. And use the shell's
'local -' feature to automate the restoration of local xtrace
option changes on function returns.

This fixes the bug with rd.live.debug where the debug_on() call
in the first getarg() call would fail to restore xtracing because
RD_DEBUG was not set. This also allows rd.live.debug to be
independent of rd.debug and eliminates 18 debug_off/debug_on calls.

This feature is present in bash and dash and not needed in mksh.
In mksh, local - (- alone) is an alias for typeset -p (synonymous
with declare -p in bash). Use local - with additional arguments to
avoid this alternate behavior in mksh.

This code works with dash since v0.5.2 (2005-09-26) and with bash
since v4.4 (2016-09-15).

Replace the nulling of debug_on()/debug_off() functions in
TEST-98-GETARG by overriding set +x in a temporary working copy of
the relevant code to aid in debugging a getarg() failure.