Always ensure /dev/{u,random} are in initramfs
In some scenarios, such as running `dracut` inside a podman/docker
container, we may not have the capabilities necessary to create
devices.

But we don't need the ability to make devices; instead, just
carry a static copy of a small cpio archive with that content
pre-created, and ensure it's the first thing in the generated
initramfs image.