feat(99~no~sh): allow a build without any command shell
Install a null link to /bin/sh in order to satisfy the executable
dependency checks and remove it in a postprocess() function.

A no-command-shell-build may be desired when the -m, --modules
option is used or when incrementally building an initramfs
with --rebuild.

This module provides an opt-out to the automatic specification
of a command shell introduced with 99~sh.